+++ kernel/src/net/net_core.rs
@@ -12,7 +12,7 @@ use crate::{
 
 use super::{
     event_poll::{EPollEventType, EventPoll},
-    socket::{TcpSocket, HANDLE_MAP, SOCKET_SET},
+    socket::{sockets::TcpSocket, HANDLE_MAP, SOCKET_SET},
 };
 
 /// The network poll function, which will be called by timer.
